Set up scripts for creating symlinks and building
The main change here is to run a pre-install script to set up
symlinks, needed because npm doesn't install symlinks even if they're
present in the git repository.

Also kept the install script to "node-gyp rebuild". It sure doesn't
seem like this should be necessary (according to the documentation
node-gyp should run if the .gyp file is present), but it seems like
it's not getting built on Heroku without this explicit script set.